fix: refer_global
