+++ src/scope/cf_scope.rs
@@ -6,7 +6,7 @@ use crate::{
 };
 use oxc::semantic::{ScopeId, SymbolId};
 use rustc_hash::FxHashSet;
-use std::rc::Rc;
+use std::{mem, rc::Rc};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
 pub enum CfScopeKind {
@@ -27,11 +27,19 @@ pub struct ExhaustiveData {
   pub deps: FxHashSet<(ScopeId, SymbolId)>,
 }
 
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+enum ReferredState {
+  Never,
+  ReferredClean,
+  ReferredDirty,
+}
+
 #[derive(Debug)]
 pub struct CfScope<'a> {
   pub kind: CfScopeKind,
   pub labels: Option<Rc<Vec<LabelEntity<'a>>>>,
   pub deps: ConsumableCollector<'a>,
+  pub referred_state: ReferredState,
   pub exited: Option<bool>,
   /// Exits that have been stopped by this scope's indeterminate state.
   /// Only available when `kind` is `If`.
@@ -50,6 +58,7 @@ impl<'a> CfScope<'a> {
       kind,
       labels,
       deps: ConsumableCollector::new(deps),
+      referred_state: ReferredState::Never,
       exited,
       blocked_exit: None,
       exhaustive_data: if kind == CfScopeKind::Exhaustive {
@@ -270,5 +279,24 @@ impl<'a> Analyzer<'a> {
     }
     self.exit_to(target_depth.unwrap());
     label_used
+  }
+
+  pub fn refer_global(&mut self) {
+    self.may_throw();
+    for depth in (0..self.scope_context.cf.stack.len()).rev() {
+      let scope = self.scope_context.cf.get_mut_from_depth(depth);
+      match scope.referred_state {
+        ReferredState::Never => {
+          scope.referred_state = ReferredState::ReferredClean;
+        }
+        ReferredState::ReferredClean => break,
+        ReferredState::ReferredDirty => {
+          scope.referred_state = ReferredState::ReferredClean;
+          let mut deps = mem::take(&mut scope.deps);
+          deps.consume_all(self);
+          break;
+        }
+      }
+    }
   }
 }
